// src/generated/caddy-encode.zod.ts
// Generated by ts-to-zod
import { z } from "zod";

export const encoderSchema = z.unknown();

export const encodingSchema = z.any();

export const precompressedSchema = z.any();

const moduleMapSchema = z.any();

const responseMatcherSchema = z.any();

export const encodeSchema = z.object({
  encodings: moduleMapSchema.optional(),
  prefer: z.array(z.string()).optional(),
  minimum_length: z.number().optional(),
  match: responseMatcherSchema.optional(),
});